Use correct type for input_offset
input_offset needs to be able to represent an arbitrary
offset inside a CattleBuffer, so it should be gulong rather
than gint.

+++ cattle/cattle-interpreter.c
@@ -77,7 +77,7 @@ struct _CattleInterpreterPrivate
 
 	gboolean             had_input;
 	CattleBuffer        *input;
-	gint                 input_offset;
+	gulong               input_offset;
 	gboolean             end_of_input_reached;
 };
